fix/#22: background 컬러 수정
기존 기본 컬러에서 White로 변경
